backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
sqlalchemy==2.0.43
asyncpg==0.31.0
alembic==1.17.2
python-dotenv==1.1.1
sqlglot==27.16.3
httpx==0.28.1
opentelemetry-api==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-exporter-otlp==1.41.0
argon2-cffi==25.1.0
itsdangerous==2.2.0
